Moved node-pre-gyp into bundled deps.
The node-pre-gyp package is necessary for our installation process in order to provide pre-built binaries. Unfortunately, there is potential for a race condition in npm 2 when resolving dependencies that could mean it is not yet installed during our 'install' script. Until we can drop Node 4 from our supported versions we must keep this as a bundled dep.